+++ src/backend/components/card/Sorcery.cpp
@@ -4,7 +4,7 @@
 
 #include "Sorcery.h"
 
-Sorcery::Sorcery(uint8_t id, const std::string &name, std::map<Color, int> cost) : Card(id, name, cost) {
+Sorcery::Sorcery(uint16_t id, const std::string &name, std::map<Color, int> cost) : Card(id, name, cost) {
 
 }
