+++ src/backend/components/deck/Graveyard.h
@@ -13,11 +13,11 @@ public :
     Graveyard();
     ~Graveyard();
 
-    uint8_t getLength() const;
+    uint16_t getLength() const;
 
     void add(std::unique_ptr<Card> card);
-    std::unique_ptr<Card> popCard(uint8_t cardId);
-    Card* getCard(uint8_t cardId);
+    std::unique_ptr<Card> popCard(uint16_t cardId);
+    Card* getCard(uint16_t cardId);
     
 
 private :
